from pc

+++ recipes/fix_nc_base.rb
@@ -2,13 +2,14 @@
 execute 'remove_repos' do
   command 'rm -f /etc/yum.repos.d/CentOS*'
   cwd '/usr/bin'
+  only_if { ::File.exist? '/etc/yum.repos.d/CentOS-Base.repo' }
   action :run
 end
 
 # make sure we have the latest RHEL versions
 execute 'update_RHEL' do
   command 'yum -y -q update'
-  action :nothing
+  action :run
 end
 
 #######################################
